UC-UBA combine tasks for each university in one script
